Fix the Fix

+++ main.py
@@ -1,8 +1,7 @@
 import hashlib, os, time
-if not os.path.exists('users/'):
-    os.mkdir('users/')
-    os.mkdir('users/guest/')
-    open('users/users.txt', 'w+').write('guest')
+if not os.path.exists('users/'): os.mkdir('users/')
+if not os.path.exists('users/guest/'): os.mkdir('users/guest/')
+if not os.path.exists('users/users.txt'): open('users/users.txt', 'w+').write('guest')
 username = input('Username: ')
 if username in open('users/users.txt', 'r').read().split('\n'):
     if username == 'guest':
